adip/requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=0.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.40.0
qdrant-client>=1.9.0
sentence-transformers>=3.0.0
umap-learn>=0.5.6
hdbscan>=0.8.33
scikit-learn>=1.5.0
kafka-python>=2.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
prometheus-client>=0.20.0
langsmith>=0.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
rank-bm25>=0.2.2
numpy>=1.26.0
pandas>=2.2.0
